Working Version. Still needsalot of work but in a spot I can start building the other stuff.

+++ Sources/CIDAFIF/CoreData/Entity+Extensions/Enity+Extensions.swift
@@ -71,7 +71,7 @@ public extension Rwy {
             return nil
         }}
     
-    class func getWithAirPortIdAndMinRunwayLength(id: String, rwyL: Int16) -> [Rwy] {
+    class func getWithAirPortIdAndMinRunwayLength(id: String, rwyL: Double) -> [Rwy] {
         var newResult: [Rwy] = []
         
         let CDFetchRequest = NSFetchRequest<Rwy>(entityName: "Rwy")
